+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add a Boolean type only JSON Schema class
 - Add a Number type only JSON Schema class
 - Add a Null type only JSON Schema class
+- Add an Object type only JSON Schema class
 
 
 ## 1.0.0-dev.9